pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["mohammad faalfard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
pandas = "^2.2.2"
lxml = "^5.2.2"
html5lib = "^1.1"
redis = "^5.0.4"
celery = "^5.4.0"
selenium = "^4.21.0"
beautifulsoup4 = "^4.12.3"
django-environ = "^0.11.2"
PyVirtualDisplay = "^3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
